+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   patterns
 - Enabling pragmas in multiline comments
   ([#116](https://github.com/JustusAdam/language-haskell/issues/116))
+- Fixed highlighting for infix constructors
+  ([#120])https://github.com/JustusAdam/language-haskell/issues/120)
 
 ## 3.0.0 - 26.04.2020